Fixed errors when creating meetings
-Glib, Jonathan

+++ src/main/java/com/vaadin/vaadin_archetype_application/Controllers.java
@@ -5,5 +5,5 @@ public class Controllers {
 	private Controllers() { } //Thank you, Java, for not having static classes /s
 	
 	public static DBProvider DatabaseConnector = (DBProvider) new DatabaseConnector();
-	public static int UserID = -1;
+	public static long UserID = -1;
 }
